# jizhangbao-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "1.0.0(jzb-api)"
    },
    "paths": {
...
        "/accrual/get": {
            "post": {
                "description": "获取指定应收应付账单",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应收应付账单"
                ],
                "summary": "获取指定应收应付账单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账单编号",
                        "name": "billId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/accrual/list/customer": {
            "post": {
                "description": "根据往来客户分组统计应收/应付数据",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询统计"
                ],
                "summary": "根据往来客户分组统计应收/应付数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本编号",
                        "name": "bookId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "账单收支类型编号 3-应收 4-应付",
                        "name": "inOut",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/accrual/list/type": {
            "post": {
                "description": "根据账单类型分组统计应收/应付数据",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询统计"
                ],
                "summary": "根据账单类型分组统计应收/应付数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本编号",
                        "name": "bookId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "账单收支类型编号 3-应收 4-应付",
                        "name": "inOut",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/bill/list/customer/month": {
            "post": {
                "description": "根据月份列出往来客户分组统计数据",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询统计"
                ],
                "summary": "根据月份列出往来客户分组统计数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本编号",
                        "name": "bookId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询月份，格式 yyyy-MM",
                        "name": "month",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "账单收支类型编号 1-收入 2-支出",
                        "name": "inOut",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bill/list/type/month": {
            "post": {
                "description": "根据月份列出账单分类分组统计数据",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "查询统计"
                ],
                "summary": "根据月份列出账单分类分组统计数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账本编号",
                        "name": "bookId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询月份，格式 yyyy-MM",
                        "name": "month",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "账单收支类型编号 1-收入 2-支出",
                        "name": "inOut",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}